Brought back CallsDialogRecording.
