+++ src/main/java/com/infobip/api/CallsApi.java
@@ -56,6 +56,7 @@ import com.infobip.model.CallsDialogPage;
 import com.infobip.model.CallsDialogPlayRequest;
 import com.infobip.model.CallsDialogRecordingPage;
 import com.infobip.model.CallsDialogRecordingRequest;
+import com.infobip.model.CallsDialogRecordingResponse;
 import com.infobip.model.CallsDialogRequest;
 import com.infobip.model.CallsDialogResponse;
 import com.infobip.model.CallsDialogSayRequest;
@@ -75,7 +76,6 @@ import com.infobip.model.CallsPlayRequest;
 import com.infobip.model.CallsPreAnswerRequest;
 import com.infobip.model.CallsPublicConferenceRecording;
 import com.infobip.model.CallsPublicCountry;
-import com.infobip.model.CallsPublicDialogRecording;
 import com.infobip.model.CallsPublicRecording;
 import com.infobip.model.CallsPublicRecordingFile;
 import com.infobip.model.CallsPublicRegion;
@@ -2828,13 +2828,13 @@ public class CallsApi {
         /**
          * Executes the deleteDialogRecordings request.
          *
-         * @return CallsPublicDialogRecording The deserialized response.
+         * @return CallsDialogRecordingResponse The deserialized response.
          * @throws ApiException If the API call fails or an error occurs during the request or response processing.
          */
-        public CallsPublicDialogRecording execute() throws ApiException {
+        public CallsDialogRecordingResponse execute() throws ApiException {
             RequestDefinition deleteDialogRecordingsDefinition = deleteDialogRecordingsDefinition(dialogId, location);
             return apiClient.execute(
-                    deleteDialogRecordingsDefinition, new TypeReference<CallsPublicDialogRecording>() {}.getType());
+                    deleteDialogRecordingsDefinition, new TypeReference<CallsDialogRecordingResponse>() {}.getType());
         }
 
         /**
@@ -2843,11 +2843,11 @@ public class CallsApi {
          * @param callback The {@link ApiCallback} to be invoked.
          * @return The {@link okhttp3.Call} associated with the API request.
          */
-        public okhttp3.Call executeAsync(ApiCallback<CallsPublicDialogRecording> callback) {
+        public okhttp3.Call executeAsync(ApiCallback<CallsDialogRecordingResponse> callback) {
             RequestDefinition deleteDialogRecordingsDefinition = deleteDialogRecordingsDefinition(dialogId, location);
             return apiClient.executeAsync(
                     deleteDialogRecordingsDefinition,
-                    new TypeReference<CallsPublicDialogRecording>() {}.getType(),
+                    new TypeReference<CallsDialogRecordingResponse>() {}.getType(),
                     callback);
         }
     }
@@ -6015,13 +6015,13 @@ public class CallsApi {
         /**
          * Executes the getDialogRecordings request.
          *
-         * @return CallsPublicDialogRecording The deserialized response.
+         * @return CallsDialogRecordingResponse The deserialized response.
          * @throws ApiException If the API call fails or an error occurs during the request or response processing.
          */
-        public CallsPublicDialogRecording execute() throws ApiException {
+        public CallsDialogRecordingResponse execute() throws ApiException {
             RequestDefinition getDialogRecordingsDefinition = getDialogRecordingsDefinition(dialogId, location);
             return apiClient.execute(
-                    getDialogRecordingsDefinition, new TypeReference<CallsPublicDialogRecording>() {}.getType());
+                    getDialogRecordingsDefinition, new TypeReference<CallsDialogRecordingResponse>() {}.getType());
         }
 
         /**
@@ -6030,11 +6030,11 @@ public class CallsApi {
          * @param callback The {@link ApiCallback} to be invoked.
          * @return The {@link okhttp3.Call} associated with the API request.
          */
-        public okhttp3.Call executeAsync(ApiCallback<CallsPublicDialogRecording> callback) {
+        public okhttp3.Call executeAsync(ApiCallback<CallsDialogRecordingResponse> callback) {
             RequestDefinition getDialogRecordingsDefinition = getDialogRecordingsDefinition(dialogId, location);
             return apiClient.executeAsync(
                     getDialogRecordingsDefinition,
-                    new TypeReference<CallsPublicDialogRecording>() {}.getType(),
+                    new TypeReference<CallsDialogRecordingResponse>() {}.getType(),
                     callback);
         }
     }
